requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
groq
chromadb
sentence-transformers
langchain-text-splitters
pymupdf
pandas
openpyxl
python-docx
pytesseract
supabase
postgrest
storage3
gotrue
realtime

